remove useless header files
