--- oneflow/core/operator/dst_subset_tick_op.h
@@ -1,43 +1,0 @@
-/*
-Copyright 2020 The OneFlow Authors. All rights reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-#ifndef ONEFLOW_CORE_OPERATOR_DST_SUBSET_TICK_OP_H_
-#define ONEFLOW_CORE_OPERATOR_DST_SUBSET_TICK_OP_H_
-
-#include "oneflow/core/operator/operator.h"
-#include "oneflow/core/graph/logical_node.h"
-
-namespace oneflow {
-
-class DstSubsetTickOp final : public Operator {
- public:
-  OF_DISALLOW_COPY_AND_MOVE(DstSubsetTickOp);
-  DstSubsetTickOp() = default;
-  ~DstSubsetTickOp() = default;
-
-  void InitFromOpConf() override;
-  Maybe<void> InferBlobDescs(std::function<BlobDesc*(const std::string&)> GetBlobDesc4BnInOp,
-                             const ParallelContext* parallel_ctx) const override;
-  LogicalNode* NewProperLogicalNode() const override;
-
- private:
-  Maybe<void> InferBatchAxis(
-      std::function<OptInt64*(const std::string&)> BatchAxis4BnInOp) const override;
-  Maybe<void> GetSbpSignatures(SbpSignatureList* sbp_sig_list) const override;
-};
-
-}  // namespace oneflow
-
-#endif  // ONEFLOW_CORE_OPERATOR_DST_SUBSET_TICK_OP_H_
